Split Offset into FixedOffset and ScalableOffset
World coordinate scaling is tricky to represent. This allows offset
providers to return an offset and indicate that it should be world-
scaled without needing to duplicate coordinate scaling logic.

It also fixes /offset set when the target player is in a scaled world.
Previously this would set the offset unscaled, but also inform the
provider of the unscaled offset, so the player's offset would change to
its scaled variant on the next change opportunity.