Restore Offset#apply
I previously moved this apply method to FixedOffset since I thought it
only made sense for an already-scaled offset to be applied. It's
actually annoying for API purposes to not be able to apply any offset
to a location. This works well since the location parameter contains a
world, so it's easy to apply a ScalableOffset to a location too.