Add offsetsAreMultiplesOfBlocks config option
This is a step towards Distant Horizons support. Distant Horizons sends
LODs that are aligned to 64-block sections, and applying a misaligned
offset to those LODs is not possible. This new config lets the server
admin control how to round offsets in case they want to have finer
control over their offsets.